Refactor SliceTransform to look more like PingCAP's
This adds a SliceTransformFns trait for the SliceTransform functions,
which can then also hold state (instead of just being pure functions).

The biggest change is probably that transform now returns a reference
slice though, instead of having to allocate a Vec for each key.

This is a breaking change, but if we wanted to upstream it we could
probably make it look at least a little more like the current API (so
only the transform_fn changes would be breaking).

See https://github.com/pingcap/rust-rocksdb/blob/master/src/slice_transform.rs